fix: changed price range type to int

+++ internal/domain/entity/product.go
@@ -5,8 +5,8 @@ type Product struct {
 	Name          string                 `json:"name"`
 	Description   string                 `json:"description"`
 	CategoryID    uint                   `json:"category_id"`
-	MinimalPrice  float64                `json:"minimal_price"`
-	MaximalPrice  float64                `json:"maximal_price"`
+	MinimalPrice  int                    `json:"minimal_price"`
+	MaximalPrice  int                    `json:"maximal_price"`
 	AverageRating float64                `json:"average_rating"`
 	CountReviews  int                    `json:"count_reviews"`
 	Attributes    map[string]interface{} `json:"product_attributes"`
